fix sorting
